created base django implementation
