+++ yt_front_end/templates/home.html
@@ -1,3 +1,4 @@
+{% load static %}
 <!DOCTYPE html>
 <head>
     <link rel="icon" type="image/x-icon" href="{% static 'ytlogo.ico' %}">
